// src/app/shared/models/VM/UserRegisterationVM.ts
export class UserRegisterationVM{
    email: string;
    password: string;
    passwordconfirm: string;
    firstName: string;
    lastName: string;
    nameArabic: string;
    phoneNumber: string;
    location: string;
    userName: string;
    defaultRole: string;
    active: boolean;
    activationPageUrl: string;
    roleName: Array<string>;
    userId: string;


}